[move-ide] Move analyzer versioning
