+++ crates/sui-move-lsp/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "move-analyzer"
+version.workspace = true
+authors = ["Mysten Labs <build@mystenlabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[dependencies]
+clap = { version = "4.1.4", features = ["derive"] }
+tokio = { workspace = true, features = ["full"] }
+
+bin-version.workspace = true
+move-lsp.workspace = true
+
